Use leveldb for peerstore
This commit adds the HYDRA_PSTORE environment variable to move from an
in-memory peerstore to a LevelDB backed one. We do this to be able to
correlate DynamoDB/Provider-Record exports with Peer-Records. We want to
be able to access and export the contents of the peer store for that.
Hence, we need to move away from an in-memory store to one that can be
accessed from other processes as well.